[Silver I] Title: 정수 삼각형, Time: 40 ms, Memory: 3980 KB -BaekjoonHub

+++ 백준/Silver/1932. 정수 삼각형/정수 삼각형.cc
@@ -1,11 +1,11 @@
 #include <bits/stdc++.h>
 using namespace std;
 
-long long arr[501][501];
-long long dp[501][501];
+int arr[501][501];
+int dp[501][501];
 
 int main() {
-	long long ans = 0;
+	int ans = 0;
 	int n; cin >> n;
 	for (int i = 0; i < n; i++) {
 		for (int j = 0; j <= i; j++) {
@@ -17,9 +17,7 @@ int main() {
 	ans = dp[0][0];
 
 	for (int i = 1, j = 1; i < n; i++, j++) {
-		// 세로줄 세팅
 		dp[i][0] = dp[i - 1][0] + arr[i][0];
-		// 오른 대각 세팅
 		dp[i][j] = dp[i - 1][j - 1] + arr[i][j];
 	}
